updated homework 15, moved user to code instead of POM
